Nano banana/nano-banana-python/pyproject.toml
[project]
name = "nano-banana-python"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["google-genai", "python-dotenv"]

[dependency-groups]
dev = ["pytest>=8.4.1"]
